Add support for enum device parameters that translate into TypeRestriction in the .knxprod file. Improved SampleDevice to reflect the new capabilities.

+++ knx-device-model/src/main/java/lu/berscheid/knx/model/KnxParameterTypeConfig.java
@@ -6,8 +6,8 @@ import lombok.Data;
 public class KnxParameterTypeConfig {
 
 	private int sizeInBit;
-	private long minInclusive;
-	private long maxInclusive;
+	private double minInclusive;
+	private double maxInclusive;
 	protected Class<?> type;
 
 	public String getName() {
